fix for prompting display logic
when re-prompting the user, do not reveal a token and then consider it a
continuous success if the user enters the passphrase correctly; this can result
in the user needing a prompt for the hidden-but-not-forgotten word, getting a
prompt every single time they enter it, and having it forgotten before they can
actually reproduce it without a prompt.

This commit was sponsored by Brian Grohe, tdsmith, Devin Prater, and
my other patrons.  If you want to join them, you can support my work
at https://glyph.im/patrons/.